+++ client/src/extension.ts
@@ -7,12 +7,7 @@ import * as Commands from './commands';
 import * as path from 'path';
 import { workspace, ExtensionContext, commands, window } from 'vscode';
 
-import {
-	LanguageClient,
-	LanguageClientOptions,
-	ServerOptions,
-	TransportKind
-} from 'vscode-languageclient';
+import { LanguageClient, LanguageClientOptions, ServerOptions, TransportKind } from 'vscode-languageclient/node';
 
 let client: LanguageClient;
 
@@ -59,7 +54,7 @@ export function activate(context: ExtensionContext) {
 	// Options to control the language client
 	let clientOptions: LanguageClientOptions = {
 		// Register the server for viz documents
-		documentSelector: ['viz', 'viz-con'],
+		documentSelector: ['viz', 'viz-con','viz4', 'viz4-con'],
 
 		synchronize: {
 			// Notify the server about file changes to '.clientrc files contained in the workspace